pyproject.toml
[project]
name = "geocode-tool"
dynamic = ["version"]
description = "Address to coordinates tool with multi-API support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [{name = "Yaemikoreal"}]
keywords = ["geocode", "geocoding", "amap", "baidu", "tianditu"]
dependencies = [
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "folium>=0.14.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.1.0"]

[project.scripts]
geocode-tool = "geocode.cli:app"

[project.urls]
Homepage = "https://github.com/Yaemikoreal/YaeLocus"
Repository = "https://github.com/Yaemikoreal/YaeLocus"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["geocode*"]

[tool.setuptools.dynamic]
version = {attr = "geocode.__version__"}